requirements.txt
alembic==1.17.2
blinker==1.9.0
brotli==1.2.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
contourpy>=1.3.3
cssselect2==0.8.0
cycler==0.12.1
dnspython==2.8.0
dominate==2.9.1
email-validator==2.3.0
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.61.1
greenlet==3.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy>=2.4,<2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.12.1
pyparsing==3.3.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.45
tinycss2==1.5.1
tinyhtml5==2.0.0
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.3
visitor==0.1.3
waitress==3.0.2
weasyprint==67.0
webencodings==0.5.1
Werkzeug==3.1.4
WTForms==3.2.1
xlrd==2.0.2
xlsxwriter==3.2.9
zopfli==0.4.0
